Extract max_std calculators to separate files
